NFDIV-4714 Refactor new Step sequences into discrete files
